add ngtouch

+++ public/scripts/controllers/singlecourseCtrl.js
@@ -145,6 +145,9 @@ app.controller('singlecourseCtrl', function($scope, $stateParams, courseService,
 		$scope.announcementsLoaded = false;
 	}
 
+	$scope.test = function() {
+		console.log("test");
+	}
 	// $scope.test = function(){
 	// 	$mdDialog.show(
 	//      	$mdDialog.alert()
@@ -177,18 +180,21 @@ app.controller('singlecourseCtrl', function($scope, $stateParams, courseService,
 	};
 
 	$scope.deleteEmail = function(email) {
-  	courseService.deleteEmail($stateParams.cid, email.itemId)
-		.then(function(res){
-			console.log("email is deleted");
-			console.log(res);
-			$scope.emailsLoaded = false;
-			$window.alert("email is deleted");
-			$scope.refreshEmails();
-		},
-		function(err){
-			console.log("Error occured : " + err);
-		});
-  	}
+		var	confirmDelete = confirm("Do you want to delete the email?");
+		if (confirmDelete == true) {
+			courseService.deleteEmail($stateParams.cid, email.itemId)
+			.then(function(res){
+				console.log("email is deleted");
+				console.log(res);
+				$scope.emailsLoaded = false;
+				$window.alert("email is deleted");
+				$scope.refreshEmails();
+			},
+			function(err){
+				console.log("Error occured : " + err);
+			});
+  		}
+	}
 
 	$scope.refreshEmails = function(){
 	courseService.getEmailbyid($stateParams.cid)
@@ -281,17 +287,20 @@ app.controller('singlecourseCtrl', function($scope, $stateParams, courseService,
 	};
 
 	$scope.deleteAnnoun = function(announcement) {
-  	courseService.deleteAnnoun($stateParams.cid, announcement.itemId)
-		.then(function(res){
-			console.log("announcement is deleted");
-			console.log(res);
-			$scope.announcementsLoaded = false;
-			$window.alert("announcement is deleted");
-			$scope.refreshAnnouns();
-		},
-		function(err){
-			console.log("Error occured : " + err);
-		});
+		var confirmDelete = confirm("Do you want to delete the announcement?");
+		if (confirmDelete == true) {
+			courseService.deleteAnnoun($stateParams.cid, announcement.itemId)
+			.then(function(res){
+				console.log("announcement is deleted");
+				console.log(res);
+				$scope.announcementsLoaded = false;
+				$window.alert("announcement is deleted");
+				$scope.refreshAnnouns();
+			},
+			function(err){
+				console.log("Error occured : " + err);
+			});
+		}
 	}
 
 	$scope.refreshAnnouns = function(){
@@ -317,7 +326,7 @@ app.controller('singlecourseCtrl', function($scope, $stateParams, courseService,
 	];
 	
 
-	function AnnounDialogController($scope, $mdDialog, $window, courseService, selectedAnnouncement, method, cid, resetLoading, refreshAnnouns) {
+	function AnnounDialogController($scope, $mdDialog, $window, $compile, courseService, selectedAnnouncement, method, cid, resetLoading, refreshAnnouns) {
 		$scope.authWrite = false;
 		$scope.authEdit = false;
 		$scope.authShow = false;
@@ -345,6 +354,8 @@ app.controller('singlecourseCtrl', function($scope, $stateParams, courseService,
 				$scope.expireEdited = tempDate;
 			}
 		}
+		// var template = angular.element($scope.currentannoun.body);
+		// $scope.currentannoun.bodyEdited = $compile(template);
 
 	  	$scope.back = function() {
 	    	$mdDialog.hide();
@@ -379,7 +390,7 @@ app.controller('singlecourseCtrl', function($scope, $stateParams, courseService,
 			});
 	  	}
 
-	  		$scope.editAnnoun = function(){
+  		$scope.editAnnoun = function(){
 	  		if ($scope.expireEdited != undefined){
 		  		var expireTime = Math.floor($scope.expireEdited.getTime()/1000);
 		  	}
@@ -408,7 +419,6 @@ app.controller('singlecourseCtrl', function($scope, $stateParams, courseService,
 			});
 	  	}
 	};
-
 
 
 function convert(array){
